feat(cli): Infer project root from absolute --why path
When an absolute path is provided to the `--why` flag, automatically
detect the project root by searching upwards for a `package.json` file.

This improves usability by allowing the tool to be run from any
directory, as long as an absolute path to a project file is given.
The `projectPath` no longer needs to be specified manually in this
scenario.